requirements.txt
pyotp==2.6.0
python-dotenv==0.20.0
robin_stocks==2.1.0
pymongo==4.1.1
python_dateutil==2.8.2
pytz==2021.1